pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fuzz-utils"
version = "0.2.0"
authors = [{ name = "Trail of Bits" }]
description = "A tool for automatically generating unit tests from Echidna and Medusa reproducers."
readme = "README.md"
license = { "text" = "AGPL-3.0" }
urls = { "Homepage" = "https://github.com/crytic/fuzz-utils" }
requires-python = ">=3.10"
dependencies = [
    "colorama>=0.4.0",
    "slither_analyzer>=0.10.0",
    "jinja2>=3.1.0",
    "eth_abi>=5.0.1",
    "eth_utils>=4.0.0",
    "eth_typing>=4.0.0",
    "pyyaml>=6.0.1"
]

[project.optional-dependencies]
lint = [
    "pylint==2.13.4",
    "black==22.3.0"
]
test = [
    "pytest",
    "solc-select>=0.1.4"
]
dev = [
    "fuzz_utils[lint,test]"
]

[project.scripts]
fuzz-utils = "fuzz_utils.main:main"

[tool.setuptools.packages.find]
    where = ["."]

[tool.black]
target-version = ["py310"]
line-length = 100
[tool.pylint.messages_control]
disable = """
unnecessary-lambda,
cyclic-import,
line-too-long,
invalid-name,
fixme,
too-many-return-statements,
too-many-ancestors,
logging-fstring-interpolation,
logging-not-lazy,
duplicate-code,
import-error,
unsubscriptable-object,
too-many-arguments,
unpacking-non-sequence
"""
[tool.mypy]
warn_incomplete_stub = true
ignore_missing_imports = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
warn_redundant_casts = true
warn_no_return = true
warn_unreachable = true